+++ backend/internal/example-custom-agent/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @cat-factory/example-custom-agent
 
+## 0.4.65
+
+### Patch Changes
+
+- Updated dependencies [afe1250]
+  - @cat-factory/kernel@0.280.0
+  - @cat-factory/agents@0.122.0
+  - @cat-factory/prompt-fragments@1.0.37
+
+## 0.4.64
+
+### Patch Changes
+
+- @cat-factory/agents@0.121.4
+- @cat-factory/kernel@0.279.3
+- @cat-factory/prompt-fragments@1.0.36
+
 ## 0.4.63
 
 ### Patch Changes